+++ lib/formatters.js
@@ -4,12 +4,11 @@ module.exports = {
             return value;
         }
         if (typeof value === 'string') {
-            try {
-                return parseInt(value, 10);
-            }
-            catch (error) {
+            const ret = parseInt(value, 10);
+            if (Number.isNaN(ret)) {
                 return defaultValue;
             }
+            return ret;
         }
         return defaultValue;
     }
